Move is_background function from chip_seq to utils
